Move *.css to `styles` folder
