+++ src/styles/CitiesSuggestionBar.css
